backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
alembic==1.14.1
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20
pydantic[email-validation]==2.10.4
email-validator==2.2.0
pydantic-settings==2.7.1
redis==5.2.1
httpx==0.28.1
beautifulsoup4==4.12.3
anthropic==0.42.0
